Make dashboars test multicluster friendly and refactor some tests so they use same suite
fix

fix

enable loadbalancer for kind

enable loadbalancer for kind only in tracing tests

deploy gateway

make ingress multiple

debug
