+++ tests/integration/telemetry/outboundtrafficpolicy/traffic_test.go
@@ -32,6 +32,7 @@ var (
 func TestMain(m *testing.M) {
 	var ist istio.Instance
 	framework.NewSuite(m).
+		RequireSingleCluster().
 		Label(label.CustomSetup).
 		Setup(istio.Setup(&ist, nil)).
 		Setup(setupPrometheus).
